Remove all warnings Add Deferred rendering Remove build artifacts Update .gitignore Fix normals Add ambient occlusion

+++ ft_vox/src/Chunk.cpp
@@ -6,7 +6,7 @@ using lazy::utils::Time;
 
 Chunk::Chunk(unsigned int seed, glm::ivec2 pos)
 {
-	_blocks = std::make_unique<octree::OctreeNode>(glm::vec3(pos.x, 32, pos.y), 32);
+	_blocks = std::make_unique<octree::OctreeNode>(glm::vec3(pos.x, 32.0f, pos.y), 32.0f);
 	_rawBlocks = std::make_unique<std::array<Block, CHUNK_SIZE * CHUNK_SIZE * CHUNK_SIZE>>();
 }
 
@@ -16,7 +16,7 @@ auto Chunk::at(std::size_t x, std::size_t y, std::size_t z) const -> Block
 		return (_void);
 	}
 
-	return _blocks->at_local(x, y, z).value_or(_void);
+	return _blocks->at_local((int)x, (int)y, (int)z).value_or(_void);
 }
 
 auto Chunk::getBlock(std::size_t x, std::size_t y, std::size_t z) const -> Block
@@ -34,5 +34,5 @@ void Chunk::setBlock(size_t x, size_t y, size_t z, Block val)
 		return ;
 	}
 	(*_rawBlocks)[x * CHUNK_SIZE * CHUNK_SIZE + y * CHUNK_SIZE + z] = val;
-	_blocks->set_local(x, y, z, val);
+	_blocks->set_local((int)x, (int)y, (int)z, val);
 }
